scripts/ingest/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
